feat: decoupling of `clingo` instance
