+++ tests/test_solver.py
@@ -3,6 +3,7 @@
 import logging
 import unittest
 
+from noqx.clingo import Config
 from noqx.manager import list_solver_metadata, load_solvers, run_solver
 from noqx.puzzle import Direction
 from noqx.rule.common import count, fill_num, unique_num
@@ -10,7 +11,6 @@ from noqx.rule.helper import fail_false, validate_direction, validate_type
 from noqx.rule.neighbor import adjacent
 from noqx.rule.shape import all_rect, all_shapes, count_shape, general_shape, get_neighbor
 from noqx.rule.variety import yaji_count
-from noqx.solution import Config
 from solver.binairo import unique_linecolor
 from solver.castle import wall_length
 from solver.heyawake import limit_border
